Fixes intro page for non protected route
