+++ app/(intro)/layout.tsx
@@ -1,0 +1,14 @@
+import { NavBar } from "@/components/Navbar";
+import { Footer } from "@/components/Footer";
+
+const MarketingLayout = ({ children }: { children: React.ReactNode }) => {
+  return (
+    <div className="h-screen bg-slate-100">
+      <NavBar />
+      <main className="pt-40 pb-20 bg-slate-100">{children}</main>
+      <Footer />
+    </div>
+  );
+};
+
+export default MarketingLayout;
