+++ app/layout.tsx
@@ -2,11 +2,6 @@ import type { Metadata } from "next";
 
 import "./globals.css";
 import { Inter } from "next/font/google";
-import { cn } from "@/lib/utils";
-
-import { NavBar } from "@/components/Navbar";
-import { Footer } from "@/components/Footer";
-import { ClerkProvider } from "@clerk/nextjs";
 
 const inter = Inter({ subsets: ["latin"] });
 
@@ -28,13 +23,7 @@ export default function RootLayout({
 }) {
   return (
     <html lang="en">
-      <body className={cn("bg-slate-100 h-full", inter.className)}>
-        <ClerkProvider>
-          <NavBar />
-          {children}
-          <Footer />
-        </ClerkProvider>
-      </body>
+      <body className={inter.className}>{children}</body>
     </html>
   );
 }
